DO NOT DOWNLOAD
Handle timeouts better.

+++ hbmame-merged-set-getter.sh
@@ -35,10 +35,10 @@ MRADIR="/media/fat/_Arcade"
 INSTALL="false"
 INIFILE="$(pwd)/update_hbmame-getter.ini"
 
-CURL_RETRY="${CURL_RETRY:---connect-timeout 15 --max-time 180 --retry 3 --retry-delay 5 --show-error}"
+CURL_RETRY="${CURL_RETRY:---connect-timeout 15 --max-time 300 --retry 3 --retry-delay 5 --show-error}"
 SSL_SECURITY_OPTION="${SSL_SECURITY_OPTION:---insecure}"
 
-EXITSTATUS=0
+rm /tmp/hbmame_getter_errors 2> /dev/null || true
 
 #####INI FILES VARS#######
 
@@ -177,13 +177,20 @@ download_hbmame_roms_from_mra() {
 
          #####CLEAN UP######
 
-         if [ ! -s "$ROMHBMAME"/"${f}" ]
-         then
+         CURL_RESULT=$?
+
+         if [[ "${CURL_RESULT}" == "28" ]] ; then
+            touch /tmp/hbmame_getter_errors
+            echo ""
+            echo "cURL error for "${f}"!"
+            echo "Try increase the max time of CURL_RETRY if the error persists."
+            rm -v "$ROMHBMAME"/"${f}" || true
+         elif [ ! -s "$ROMHBMAME"/"${f}" ] ; then
             echo ""
             echo "0 byte file found for "${f}"!"
             echo "This happens when the file is missing or unavailable from the download source."
             rm -v "${ROMHBMAME}"/"${f}"
-            EXITSTATUS=1
+            touch /tmp/hbmame_getter_errors
          fi
 
          echo
@@ -292,18 +299,22 @@ hbmame_getter_optimized() {
       download_hbmame_roms_from_mra "${i}"
    done
 
-   echo "${HBMAME_GETTER_VERSION}" > "${LAST_RUN_PATH}"
-   echo "${INI_DATE}" >> "${LAST_RUN_PATH}"
-   echo "${MRA_DATE}" >> "${LAST_RUN_PATH}"
+   if [ ! -f /tmp/hbmame_getter_errors ] ; then
+      echo "${HBMAME_GETTER_VERSION}" > "${LAST_RUN_PATH}"
+      echo "${INI_DATE}" >> "${LAST_RUN_PATH}"
+      echo "${MRA_DATE}" >> "${LAST_RUN_PATH}"
+   fi
 }
 
 if [ ${#} -eq 2 ] && [ ${1} == "--input-file" ] ; then
+
    MRA_INPUT="${2:-}"
    if [ ! -f ${MRA_INPUT} ] ; then
       echo "Option --input-file selected, but file '${MRA_INPUT}' does not exist."
       echo "Usage: ./${0} --input-file file"
       exit 1
    fi
+
    echo ""
    echo "$(wc -l ${MRA_INPUT} | awk '{print $1}') arguments provided, this script expects them to be valid .mra files."
    echo ""
@@ -337,19 +348,17 @@ fi
 rm /tmp/hbmame.getter.zip.file
 rm /tmp/hbmame.getter.mra.file
 
-echo
-echo "SUCCESS!"
-echo
-
-######INFO#####
-if [ -d "/media/fat/_Arcade/hbmame" ] ; then
+if [ ! -f /tmp/hbmame_getter_errors ] ; then
    echo
-   echo "INFO: As of 6/11/2020 the default directory has been changed to /media/fat/games/hbmame"
-   echo "INFO: Please move all roms from /media/fat/_Arcade/hbmame/* to /media/fat/games/hbmame/"
-   echo "INFO: You may still set a custom ROMHBMAME path in update_hbmame-getter.ini if needed"
+   echo "SUCCESS!"
+   echo
+   exit 0
+else
+   echo
+   echo "Some error happened. Try again later!"
+   echo
+   exit 1
 fi
-
-exit ${EXITSTATUS}
 
 #####MERGED .220 LIST######
 ##HBMAME .220 LIST
